Added a few more validation considerations, added and tested an alt PricingRuleset to ensure new rulesets could be added, decided to change all my doubles back to ints as this matched the spec closer. removing last warnings for nulls - 15 tests passing

+++ src/PricingRules/PricingRules.cs
@@ -3,9 +3,9 @@
 public class PricingRule
 {
     public required string Sku { get ; set; }
-    public required double price { get ; set; }
+    public required int price { get ; set; }
     public int? SpecialAmount { get ; set; }
-    public double? SpecialPrice {get; set;} //both nullable as not every item has a Deal
+    public int? SpecialPrice {get; set;} //both nullable as not every item has a Deal
 
 }
 public class PricingRulesRepo: IPricingRulesRepo
