+++ src/interfaces/ICheckout.cs
@@ -4,5 +4,5 @@ using CheckoutKata;
 public interface ICheckout
 {
     void Scan(string item);
-    double GetTotalPrice();
+    int GetTotalPrice();
 }
